+++ src/util/figure.test.ts
@@ -72,9 +72,9 @@ describe('Unit: util/figure', () => {
     it('fixed', () => {
       const source = parse([
         '~~~figure [:fig-1]\n!https://host\n~~~',
-        '~~~figure [:fig-1.1]\n!https://host\n~~~',
+        '~~~figure [:fig-a-0.0]\n!https://host\n~~~',
         '~~~figure [:$-2.1]\n$$\nLaTeX\n$$\n~~~',
-        '[:fig-1.1]',
+        '[:$-2.1]',
       ].join('\n\n'));
       for (let i = 0; i < 3; ++i) {
         figure(source);
@@ -82,9 +82,9 @@ describe('Unit: util/figure', () => {
           [...source.children].map(el => el.outerHTML),
           [
             '<figure class="label:fig-1" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
-            '<figure class="label:fig-1.1" data-group="fig" data-index="1.1" id="label:fig-1.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
+            '<figure class="label:fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-1.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
             '<figure class="label:$-2.1" data-group="$" data-index="2.1" id="label:$-2.1"><div class="figcontent"><div class="math notranslate">$$\nLaTeX\n$$</div></div><span class="figindex">(2.1)</span><figcaption></figcaption></figure>',
-            '<p><a href="#label:fig-1.1" rel="noopener" class="label:fig-1.1">Fig. 1.1</a></p>',
+            '<p><a href="#label:$-2.1" rel="noopener" class="label:$-2.1">(2.1)</a></p>',
           ]);
       }
     });
@@ -107,16 +107,18 @@ describe('Unit: util/figure', () => {
 
     it('separation', () => {
       const source = parse([
-        '~~~figure [:fig-a]\n!https://host\n~~~',
+        '!>> ~~~figure [:fig-1]\n!https://host\n~~~\n> ~~~figure [:fig-a]\n!https://host\n~~~',
         '~~~~example/markdown\n~~~figure [:fig-a]\n!https://host\n~~~\n~~~~',
+        '~~~figure [:fig-a]\n!https://host\n~~~',
       ].join('\n\n'));
       for (let i = 0; i < 3; ++i) {
         figure(source);
         assert.deepStrictEqual(
           [...source.children].map(el => el.outerHTML),
           [
-            '<figure class="label:fig-a" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
+            '<blockquote><blockquote><figure class="label:fig-1" data-group="fig" data-index="1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure></blockquote><figure class="label:fig-0" data-group="fig" data-index="0"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 0.</span><figcaption></figcaption></figure></blockquote>',
             '<aside class="example" data-type="markdown"><pre>~~~figure [:fig-a]\n!https://host\n~~~</pre><div><figure class="label:fig-a" data-group="fig" data-index="1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure></div><ol></ol><ol></ol></aside>',
+            '<figure class="label:fig-a" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
           ]);
       }
     });