+++ src/util/figure.ts
@@ -5,24 +5,31 @@ import { define } from 'typed-dom';
 
 export function figure(source: DocumentFragment | HTMLElement): void {
   const groups = new Map<string, HTMLElement[]>();
-  const exclusions = new Set<Element>(source.querySelectorAll('.example figure'));
+  const exclusions = new Set<Element>();
   return void source.querySelectorAll<HTMLElement>('figure[class^="label:"]')
-    .forEach(figure => {
-      if (exclusions.has(figure)) return;
-      const label = figure.className;
-      const group = figure.getAttribute('data-group')!;
-      void groups.set(group, concat(groups.get(group) || [], [figure]));
+    .forEach(fig => {
+      if (fig.matches('.example figure')) return;
+      if (fig.parentElement !== source && fig.parentElement instanceof HTMLQuoteElement) {
+        return exclusions.has(fig.parentElement)
+          ? undefined
+          : void exclusions.add(fig.parentElement) ||
+            void figure(fig.parentElement);
+      }
+      const label = fig.className;
+      const group = fig.getAttribute('data-group')!;
+      void groups.set(group, concat(groups.get(group) || [], [fig]));
       assert(groups.has(group));
       const idx = index(label, groups.get(group)!);
-      void figure.setAttribute('data-index', idx);
-      void figure.setAttribute('id', `${label.split('-', 1)[0]}-${idx}`);
-      const figindex = figure.lastElementChild!.previousElementSibling!;
+      void fig.setAttribute('data-index', idx);
+      const figindex = fig.lastElementChild!.previousElementSibling!;
       assert(figindex.matches('.figindex'));
       void define(figindex, group === '$' ? `(${idx})` : `${capitalize(group)}. ${idx}.`);
+      if (fig.closest('blockquote')) return;
+      void fig.setAttribute('id', `${label.split('-', 1)[0]}-${idx}`);
       const query = isGroup(label) ? label.split('-').slice(0, -1).join('-') : label;
       void source.querySelectorAll(`a.${query.replace(/[:$.]/g, '\\$&')}`)
         .forEach(ref =>
-          void define(ref, { href: `#${figure.id}` }, figindex.textContent!.replace(/[.]$/, '')));
+          void define(ref, { href: `#${fig.id}` }, figindex.textContent!.replace(/[.]$/, '')));
     });
 }